New JSON format
Now, the JSON file is smarter and with a few more meta-informations, like geolocation of the host city. This will make easier to use the JSON as an static API for FEMUG infos.

this format is better, and also introduce the fields in english.
first commit with the globalisation intent :)